back-end drcare/requirements.txt
Django==3.2.14
django-cors-headers==3.13.0
djangorestframework==3.13.1
sklearn==0.0
experta==1.9.4
h5py==2.10.0
Pillow==9.1.1
protobuf==3.19.1
keras==2.3.1
pytesseract==0.3.9
opencv-python==4.6.0.66
spacy==3.4.0
numpy==1.21.6
whitenoise==6.2.0
gunicorn==20.1.0
